+++ docs/setup/changelog.rst
@@ -16,6 +16,7 @@ Upcoming Changes (main branch)
 
 - Added ``array.copy()`` and ``nd.copy(array)`` functions.
 - Added ``array.copy()`` and ``nd.copy(array)`` functions.
+- Added ``positive`` and ``negative`` functions.
 
 **Changed**